+++ x/stake/types_test.go
@@ -1,3 +1,0 @@
-package stake
-
-// XXX test global state functions, validator exchange rate functions etc.